HW_6 Release

+++ Semester_1/Fq1jjeR/HW_6/3_Task.cpp
@@ -11,8 +11,8 @@ int main() {
     cout << "Введите количевство M (столбцов): ";
     cin >> M;
 
-    int count_N = N * (N + 1) / 2;
-    int count_M = M * (M + 1) / 2;
+    long long count_N = N * (N + 1) / 2;
+    long long count_M = M * (M + 1) / 2;
 
     cout << "Количество прямоугольников: " << count_N * count_M << endl;
